backend/requirements.txt
fastapi==0.124.0
uvicorn[standard]==0.34.2
python-dotenv==1.2.1
openai==1.82.0
cohere==5.20.0
qdrant-client==1.16.1
google-generativeai==0.8.5
pydantic==2.11.5
pytest>=8.2.0
pytest-asyncio>=0.21.0
requests==2.32.3
pydantic-settings==2.9.1
slowapi==0.1.9
redis==7.1.0